Fix IPv4 octet and mask length overflow (#57)
The `dig` function in `Data.IP.Addr` uses an `Int` accumulator that
can overflow while parsing.  It is used to parse IPv4 octets as well
as IPv4 and IPv6 mask lengths.  This change replaces the `dig`
function with two special-purpose functions:

The `octet` function parses IPv4 octets.  It checks the maximum bound
while parsing so that overflow is not possible.

The `maskLen` function parses mask lengths.  The maximum bound is
passed as an argument (32 for IPv4 and 128 for IPv6), and it checks
the maximum bound while parsing so that overflow is not possible.  The
use of `option` is refactored to only check for the `/` character, so
that an invalid mask length after the `/` character can cause a parse
failure.

Note that these functions include the following performance
optimizations:

* The `ord` function is used instead of `digitToInt`, as the range of
  characters is already checked.
* The parameters to the `go` helper function are given (bang)
  strictness annotations to help GHC determine that the `Char` and
  `Int` values can be unboxed.

An unfortunate side effect of this change is that some error messages
change.